file reading

+++ src/com/thehe/WHGRL/utils/Vector.java
@@ -2,15 +2,15 @@ package com.thehe.WHGRL.utils;
 
 public class Vector {
 	
-	public double x;
-	public double y;
+	public int x;
+	public int y;
 	
 	public Vector() {
 		this.x = 0;
 		this.y = 0;
 	}
 	
-	public Vector(double x, double y) {
+	public Vector(int x, int y) {
 		this.x = x;
 		this.y = y;
 	}
